Mamba dis

+++ src/dissimilarity_learning.py
@@ -1,50 +1,83 @@
 import os
 from subprocess import call
+import argparse
 
-def generate_and_submit_scripts():
+
+def generate_and_submit_scripts(args: argparse.Namespace):
     script_template = """#!/bin/bash
 #SBATCH --nodes=1
 #SBATCH --time=24:00:00
-#SBATCH --job-name=dissimilarity_over_learning
-#SBATCH --gres=gpu:1
-#SBATCH --ntasks-per-node=1
-#SBATCH --cpus-per-task=10 
+#SBATCH --job-name=dissimilarity_over_learning_{taskset}_{group1}_{group2}
+#SBATCH --gres=gpu:8
 #SBATCH --mail-type=ALL
 #SBATCH --mail-user=oxfd2547@ox.ac.uk
 
 module load cuda/11.2
 module load pytorch/1.9.0
-module load python/anaconda3
+module load python/miniconda3
 
-source activate dsa
-poetry install
+source activate repal3
 
-(poetry run python -m src.dissimilarity_over_learning_per_group --group1 {group1} --group2 {group2}) & 
+
+(python -m src.dissimilarity_over_learning_per_group --taskset {taskset} --group1 {group1} --group2 {group2}) & 
 
 # PID of the application
 APP_PID=$!
 
-# Monitor GPU status every 300 seconds (5 minutes) until the application finishes
+# Log file path
+LOG_FILE=gpu_usage/learning_{taskset}_{group1}__{group2}_gpu_usage.log
+
+# Monitor GPU status every 600 seconds (5 minutes) until the application finishes
 while kill -0 $APP_PID 2>/dev/null; do
-    echo "$(date '+%Y-%m-%d %H:%M:%S') - Checking GPU status during the application run:" >> gpu_usage/learning_{group1}_{group2}_gpu_usage.log
-    nvidia-smi >> gpu_usage/learning_{group1}_{group2}_gpu_usage.log  # Append output to log file
-    sleep 300 
+    {{
+        echo "$(date '+%Y-%m-%d %H:%M:%S') - Checking GPU status during the application run:"
+        nvidia-smi
+    }} >> "$LOG_FILE"  # Append output to log file
+    sleep 600 
 done
 
 wait $APP_PID
-
 """
-    groups_to_compare_over_learning = [("pretrain_frozen","pretrain_unfrozen"),("pretrain_frozen","master"),("pretrain_unfrozen","master")]
+    groups = [
+        "untrained",
+        "master_frozen",
+        "master",
+        "pretrain_basic_frozen",
+        "pretrain_anti_frozen",
+        "pretrain_delay_frozen",
+        "pretrain_basic_anti_frozen",
+        "pretrain_basic_delay_frozen",
+        "pretrain_frozen",
+        "pretrain_unfrozen",
+    ]
 
-    for group in groups_to_compare_over_learning:
-        script_content = script_template.format(group1=group[0], group2=group[1])
-        script_filename = f"sbatch/dissimilarities_over_learning/{group[0]}_{group[1]}_script.sh"
+    for i in range(len(groups)):
+        for j in range(len(groups)):
+            group = [groups[i], groups[j]]
+            folder_npz_name_1 = f"data/dissimilarities_over_learning/{args.taskset}/{group[0]}_{group[1]}"
+            folder_npz_name_2 = f"data/dissimilarities_over_learning/{args.taskset}/{group[1]}_{group[0]}"
+            if not os.path.exists(folder_npz_name_1) and not os.path.exists(
+                folder_npz_name_2
+            ):
+                script_content = script_template.format(
+                    taskset=args.taskset, group1=group[0], group2=group[1]
+                )
+                script_filename = f"sbatch/dissimilarities_over_learning/{args.taskset}/{group[0]}_{group[1]}_script.sh"
 
-        with open(script_filename, 'w') as script_file:
-            script_file.write(script_content)
+                with open(script_filename, "w") as script_file:
+                    script_file.write(script_content)
 
-        # Submit the job to the cluster
-        call(f"sbatch {script_filename}", shell=True)
+                # Submit the job to the cluster
+                call(f"sbatch {script_filename}", shell=True)
+
 
 if __name__ == "__main__":
-    generate_and_submit_scripts()
+    parser = argparse.ArgumentParser(description="Train the model")
+    parser.add_argument(
+        "--taskset",
+        type=str,
+        default="PDM",
+        help="taskset to compare dissimilarities on",
+    )
+    args = parser.parse_args()
+    generate_and_submit_scripts(args)
